requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.1.2
dj-static==0.0.6
Django==3.1.1
django-allauth==0.43.0
django-cors-headers==3.5.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-redis==4.12.1
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.1
Faker==4.14.0
gunicorn==20.0.4
idna==2.10
mysql-client==0.0.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.11
whitenoise==5.2.0
